Fix errors in LLVM 4.1 (Xcode 4.5)

+++ CoreLib/Extensions/NSData+PivotalCore.m
@@ -29,7 +29,7 @@
     NSMutableString *hexString = [NSMutableString stringWithCapacity:(dataLength * 2)];
 
     for (int i = 0; i < dataLength; ++i) {
-        [hexString appendString:[NSString stringWithFormat:@"%02x", (unsigned long)dataBuffer[i]]];
+        [hexString appendString:[NSString stringWithFormat:@"%02x", dataBuffer[i]]];
     }
 
     return [NSString stringWithString:hexString];
